+++ back-end/services/analytics-service/src/main/java/com/project/mentorship/service/analytics/domain/Statistics.java
@@ -1,6 +1,5 @@
 package com.project.mentorship.service.analytics.domain;
 
-import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 import java.util.UUID;
 import lombok.AllArgsConstructor;
@@ -15,7 +14,7 @@ public class Statistics {
 	private UUID id;
 	private OffsetDateTime date;
 	private Integer totalReservations;
-	private BigDecimal totalRevenue;
+	private Double totalRevenue;
 	private OffsetDateTime createdAt;
 
 }
