+++ back-end/services/analytics-service/src/test/java/com/project/mentorship/service/analytics/api/StatisticsControllerTest.java
@@ -6,7 +6,6 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.project.mentorship.service.analytics.AnalyticsServiceApplication;
 import com.project.mentorship.service.analytics.api.dto.StatisticsDto;
-import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
 import org.junit.jupiter.api.Test;
@@ -30,7 +29,7 @@ class StatisticsControllerIT {
 	void createStatistics_shouldReturn201Created_whenRequestIsValid() throws Exception {
 		// Given
 		StatisticsDto dto = new StatisticsDto(null, OffsetDateTime.of(2025, 1, 1, 0, 0, 0, 0, ZoneOffset.UTC), 10,
-				BigDecimal.valueOf(250), null);
+				Double.valueOf(250.0), null);
 
 		String requestBody = objectMapper.writeValueAsString(dto);
